requirements.txt
asgiref==3.4.1
Brotli==1.0.9
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.2
click==8.0.1
colorama==0.4.4
cycler==0.10.0
dash==1.20.0
dash-bootstrap-components==0.12.2
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
Django==3.2.6
djangorestframework==3.12.4
Flask==2.0.1
Flask-Compress==1.10.1
future==0.18.2
google-auth==1.33.0
google-auth-oauthlib==0.4.4
gspread==3.7.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.0
oauthlib==3.1.1
pandas==1.2.5
Pillow==8.3.1
plotly==5.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.0
six==1.16.0
sqlparse==0.4.1
tenacity==7.0.0
urllib3==1.26.6
Werkzeug==2.0.1
